+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.9.0] - 2025-09-03
+
+### Added
+
+- Newsletter subscription modal
+- WOETH pool support
+- Custom token pricing support
+
+### Changed
+
+- Removed fees when using native token
+
+### Fixed
+
+- Fixed SDK issues with deposits
+
 ## [1.8.0] - 2025-08-21
 
 ### Changed